fix: put a limit on mainIndex based on the length of proof
In some cases start and end block numbers might be way bigger than length of the tree. For example when it is a first checkpoint when starting block number is 0 and end can be any. In this case we need to put an upper bound on index to meet checkMembership criteria.